# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/instance/profile:
        get:
            tags:
                - InstanceService
            description: Gets the instance profile.
            operationId: InstanceService_GetInstanceProfile
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InstanceProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        InstanceProfile:
            type: object
            properties:
                owner:
                    type: string
                    description: |-
                        The name of instance owner.
                         Format: users/{user}
                version:
                    type: string
                    description: Version is the current version of instance.
                demo:
                    type: boolean
                    description: Demo indicates if the instance is in demo mode.
                instanceUrl:
                    type: string
                    description: Instance URL is the URL of the instance.
            description: Instance profile message containing basic instance information.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...